+++ cassandra-four-zero-bridge/src/main/java/org/apache/cassandra/spark/reader/CompressedRawInputStream.java
@@ -34,8 +34,8 @@ import org.slf4j.LoggerFactory;
 import org.apache.cassandra.spark.data.SSTable;
 import org.apache.cassandra.spark.reader.common.AbstractCompressionMetadata;
 import org.apache.cassandra.spark.reader.common.ChunkCorruptException;
-import org.apache.cassandra.spark.reader.common.RawInputStream;
-import org.apache.cassandra.spark.stats.Stats;
+import org.apache.cassandra.analytics.reader.common.RawInputStream;
+import org.apache.cassandra.analytics.stats.Stats;
 import org.apache.cassandra.utils.ByteBufferUtil;
 import org.apache.cassandra.utils.ChecksumType;
 import org.jetbrains.annotations.Nullable;
